Change CreateTransaction subaccount to u32, to match everywhere else

+++ subprojects/gdk_rust/gdk_common/src/model.rs
@@ -126,7 +126,7 @@ pub struct CreateTransaction {
 pub struct GetTransactionsOpt {
     pub first: usize,
     pub count: usize,
-    pub subaccount: usize,
+    pub subaccount: u32,
     pub num_confs: Option<usize>,
 }
